Move OperatorHandle into separate module
Working towards only having the public operator interface in
oremda.operator
