+++ oremda/__init__.py
@@ -1,6 +1,7 @@
 from .plasma_client import PlasmaArray, PlasmaClient
-from .operator import Operator, OperatorHandle, operator
+from .operator import Operator, operator
 from .pipeline import Pipeline
+from .pipeline.operator import OperatorHandle
 
 __all__ = [
     "Operator",
